% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optFunctions.R
\name{blockNorms}
\alias{blockNorms}
\title{Get Frobenius norms of submatrices with optional weights}
\usage{
blockNorms(M, id, W = NULL)
}
\arguments{
\item{M}{square matrix of interest}

\item{id}{vector grouping elements of M}

\item{W}{optional weights (square matrix, dimensions equal to unique id}
}
\value{
matrix of Frobenius norms of submatrices
}
\description{
Given a square matrix \code{M} and a vector \code{id} that distinguishes sections of \code{M},
return a matrix with the Frobenius norm of the specified submatrices.
Optionally, provide a square weight matrix \code{W} with number of rows and columns
equal to the unique entries in id that multiplies these norms.
}

