% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findPerturbations.R
\name{findPerturbations}
\alias{findPerturbations}
\title{Get node-wise test statistics}
\usage{
findPerturbations(Y, Omega, Sigma, id, sequential = FALSE)
}
\arguments{
\item{Y}{matrix or data frame (rows=subjects, columns=variables)}

\item{Omega}{precision matrix}

\item{Sigma}{covariance matrix}

\item{id}{vector mapping variables to nodes}

\item{sequential}{boolean, whether or not to perform sequential adjustments}
}
\value{
test statistics  for perturbations at each individual node
}
\description{
Given a vector \code{id} indicating node membership, test for perturbations
at each node as described by Griffin et al.  This function performs either
the single-target test (\code{sequential=FALSE}, default) or the
multi-target adjusted procedure  (\code{sequential=TRUE})
}
\references{
Griffin, P.J., Johnson, W.E., and Kolaczyk, E.K. Detection of multiple
perturbations in multi-omics biological networks' (under review)
}

