% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optFunctions.R
\name{getDual}
\alias{getDual}
\title{Get value of dual function (called by maLasso1)}
\usage{
getDual(Sigma)
}
\arguments{
\item{Sigma}{current estimated covariance}
}
\value{
value of dual function for optimization problem
}
\description{
Get value of dual function (called by maLasso1)
}

