% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optFunctions.R
\name{getEBIC}
\alias{getEBIC}
\title{Get extended BIC of a given network configuration}
\usage{
getEBIC(S, n, Omega, id, gamma = 0.5)
}
\arguments{
\item{S}{sample covariance matrix}

\item{n}{number of samples}

\item{Omega}{estimated precision}

\item{id}{vector grouping elements of S, Omega}

\item{gamma}{EBIC parameter (default 0.5)}
}
\value{
extended Bayesian information criterion for model implied by Omega
}
\description{
Get extended BIC of a given network configuration
}

