% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optFunctions.R
\name{getLambdaRange}
\alias{getLambdaRange}
\title{Generate lambdas to try}
\usage{
getLambdaRange(S, id, length.out = 10)
}
\arguments{
\item{S}{sample covariance}

\item{id}{vector of node identifiers}

\item{length.out}{number of lambda parameters to return}
}
\value{
vector of \code{length.out} equally spaced lambdas
}
\description{
Generate penalty parameters spanning a range of 1 to
\code{length(unique(id))} submatrix blocks.  This function is provided to
establish a range of reasonable penalty parameters for optimization
according to the algorithm of Kolar et al (2014).
}
\examples{
Y <- matrix(rnorm(120), nrow=20, ncol=6)
S <- crossprod(Y)
id <- rep(1:3, each=2)
getLambdaRange(S, id, 5)
}
\references{
Kolar, M., Liu, H., and Xing, E. P. (2014). Graph estimation from
multi-attribute data. The Journal of Machine Learning Research 15, 1713-1750.
}

