% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optFunctions.R
\name{getLambdas}
\alias{getLambdas}
\title{Generate lambdas to try}
\usage{
getLambdas(S, id, length.out = 10)
}
\arguments{
\item{S}{sample covariance}

\item{id}{vector of node identifiers}

\item{length.out}{number of lambda parameters to return}
}
\value{
vector of length.out equally spaced lambdas
}
\description{
Generate penalty parameters spanning a range of 1 to n.nodes submatrix blocks
}

