% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optFunctions.R
\name{getPrimal}
\alias{getPrimal}
\title{Get value of primal function (called by maLasso1)}
\usage{
getPrimal(Omega, S, W, id, lambda)
}
\arguments{
\item{Omega}{current estimated precision}

\item{S}{sample covariance matrix}

\item{W}{weight matrix}

\item{id}{vector grouping elements of S, Omega}

\item{lambda}{penalty parameter}
}
\value{
value of primal function for optimization problem
}
\description{
Get value of primal function (called by maLasso1)
}

