% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findPerturbations.R
\name{groupCondLRTs}
\alias{groupCondLRTs}
\title{Sequentially test for perturbations at a series of locations}
\usage{
groupCondLRTs(Y, Omega, Sigma, perturb.mat, ret = "stat")
}
\arguments{
\item{Y}{matrix or data frame (rows=subjects, columns=variables)}

\item{Omega}{precision matrix}

\item{Sigma}{covariance matrix}

\item{perturb.mat}{matrix indicating sets of which variables to test for perturbations}

\item{ret}{return statistics ('stat') or p-values ('pval')}
}
\value{
test statistics for perturbations at each of perturb.mat rows
}
\description{
Sequentially test for perturbations at a series of locations
}

