% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findPerturbations.R
\name{groupLRT}
\alias{groupLRT}
\title{Test for a perturbation at a specific location}
\usage{
groupLRT(Y, Omega, Sigma, perturb.vec, ret = "stat")
}
\arguments{
\item{Y}{matrix or data frame (rows=subjects, columns=variables)}

\item{Omega}{precision matrix}

\item{Sigma}{covariance matrix}

\item{perturb.vec}{vector indicating which variables to test for perturbations}

\item{ret}{value to return ('stat' or 'pval')}
}
\value{
test statistic for a perturbation at perturb.vec=TRUE locations only
}
\description{
Test for a perturbation at a specific location
}

