% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findPerturbations.R
\name{groupLRTs}
\alias{groupLRTs}
\title{Test for perturbations at a series of locations}
\usage{
groupLRTs(Y, Omega, Sigma, perturb.mat, ret = "stat")
}
\arguments{
\item{Y}{matrix or data frame (rows=subjects, columns=variables)}

\item{Omega}{precision matrix}

\item{Sigma}{covariance matrix}

\item{perturb.mat}{matrix indicating sets of which variables to test for perturbations}

\item{ret}{value to return ('stat' or 'pval')}
}
\value{
test statistics for perturbations at each of perturb.mat rows
}
\description{
Test for perturbations at a series of locations
}

