% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{isComputationallyPD}
\alias{isComputationallyPD}
\title{Check if a matrix is computationally positive-definite}
\usage{
isComputationallyPD(M)
}
\arguments{
\item{M}{matrix of interest}
}
\value{
TRUE if matrix is computationally positive-definite
}
\description{
Check if a matrix is computationally positive-definite
}

