% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maLasso.R
\name{maLasso}
\alias{maLasso}
\title{Multi-attribute network estimation}
\usage{
maLasso(S, n, id, lambda, W = NULL, update = 100, max.gap = 0.5,
  max.iter = 100, min.t = .Machine$double.eps)
}
\arguments{
\item{S}{sample covariance matrix}

\item{n}{number of samples}

\item{id}{vector of node identifiers}

\item{lambda}{tuning parameter for penalty}

\item{W}{optional penalty weight matrix (same dimensions as S)}

\item{update}{how often to print updates}

\item{max.gap}{maximum allowable primal-dual gap}

\item{max.iter}{maximum number of iterations to complete (overriedes max.gap)}

\item{min.t}{minimum step size (overrides max.gap, max.iter)}
}
\value{
list of precision, covariance, optimization status, lambda, and number components
}
\description{
Estimates block precisions and covariances for a multi-attribute network
based on a Gaussian graphical model with zero mean vector.  This is according
to the method described by Kolar et al (2015).
}
\references{
Kolar, M., Liu, H., and Xing, E. P. (2014). Graph estimation from
multi-attribute data. The Journal of Machine Learning Research 15, 1713–1750.
}

