% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lassoWraps.R
\name{multiAttSelect}
\alias{multiAttSelect}
\title{Wrapper function for multi-attribute network estimation plus selection via lasso}
\usage{
multiAttSelect(S, n, id, lambda.range, W = NULL, mode = 1, update = 100,
  max.gap = 0.5, max.iter = 100, min.t = .Machine$double.eps,
  method = "BIC", Theta.true = NULL, plot = NULL, gamma = 0.5)
}
\arguments{
\item{S}{sample covariance matrix}

\item{n}{number of samples}

\item{id}{vector assigning variables to nodes}

\item{lambda.range}{vector of lambdas to try}

\item{W}{optional weight matrix}

\item{mode}{estimation mode. Can be 1 (multiattribute; default),
2 (unstructured together), or 3 (separately; only valid for 2-attribute data)}

\item{update}{how often to print updates in optimization}

\item{max.gap}{maximum allowable primal/dual gap}

\item{max.iter}{maximum number of iterations to optimize (overrides max.gap)}

\item{min.t}{minimum step size (overrides max.gap, max.iter}

\item{method}{how to select the best model (\code{'EBIC'}, \code{'BIC'},
\code{'AIC'}, or \code{'hamming'})}

\item{Theta.true}{true underlying graph (\code{'hamming'} method only)}

\item{plot}{boolean, whether or not to make diagnostic plot}

\item{gamma}{gamma parameter for EBIC method (default 0.5)}
}
\value{
list of precision, covariance, optimization status, lambda, and
number components
}
\description{
Performs estimation of a zero-mean multi-attribute Gaussian graphical model
as described by Kolar et al (2014) when \code{mode=1} with a selection
procedure to determine optimal \code{lambda}.  Selection may be performed
according to extended BIC, BIC, AIC, or Hamming distance to true graph.
The true graph \code{Theta} is only required if Hamming distance is being
used for the selection procedure.
}
\details{
If \code{mode=2}, the \code{id} argument is essentially disregarded, and all
rows/columns in \code{S} are treated as if they represent individual nodes.
If \code{mode=3}, the optimization of \code{mode=2} is performed separately
for each attribute type, and a matrix with 0 on all cross-attribute entries
is returned.
}
\examples{
library(mvtnorm)
id <- rep(1:3, each=2)
Omega <- matrix(0, nrow=6, ncol=6)
Omega[1:4,1:4] <- -1
diag(Omega) <- 6
Sigma <- solve(Omega)
n <- 1000
Y <- rmvnorm(n, sigma=Sigma)
S <- crossprod(Y)

lambdas <- getLambdaRange(S, id, 3)
result <- multiAttSelect(S, n, id, lambdas)
}
\references{
Kolar, M., Liu, H., and Xing, E. P. (2014). Graph estimation from
multi-attribute data. The Journal of Machine Learning Research 15, 1713-1750.

Chen, J. and Chen, Z. (2008). Extended Bayesian information criteria for
model selection with large model spaces. Biometrika 95, 759-771.
}

