% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findPerturbations.R
\name{perturbNodeTests}
\alias{perturbNodeTests}
\title{Get node-wise test statistics}
\usage{
perturbNodeTests(Y, Omega, Sigma, id, sequential = FALSE,
  return.value = "stat")
}
\arguments{
\item{Y}{matrix or data frame (rows=subjects, columns=variables)}

\item{Omega}{precision matrix}

\item{Sigma}{covariance matrix}

\item{id}{vector mapping variables to nodes}

\item{sequential}{boolean, whether or not to perform sequential adjustments}

\item{return.value}{return statistics (\code{'stat'}) or p-values (\code{'pval'})}
}
\value{
test statistics or p-values for perturbations at each individual node
}
\description{
Given a vector \code{id} indicating node membership, test for perturbations
at each node as described by Griffin et al.  This function performs either
the single-target test (\code{sequential=FALSE}, default) or the
multi-target adjusted procedure  (\code{sequential=TRUE})
}
\examples{
library(mvtnorm)
Omega <- matrix(0, nrow=6, ncol=6)
Omega[1:4,1:4] <- -1
diag(Omega) <- 6
Sigma <- solve(Omega)
mu <- c(1,1,0,0,0,0)
Y <- rmvnorm(20, mean=Sigma \%*\% mu, sigma=Sigma)
id <- c(1,1,2,2,3,3)
perturbNodeTests(Y, Omega, Sigma, id, sequential=FALSE) # unadjusted
perturbNodeTests(Y, Omega, Sigma, id, sequential=TRUE) # sequential
}
\references{
Griffin, P.J., Johnson, W.E., and Kolaczyk, E.K. Detection of multiple
perturbations in multi-omics biological networks (under review)
}

