% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lassoWraps.R
\name{runLasso}
\alias{runLasso}
\title{Wrapper function for multi-attribute network estimation via lasso}
\usage{
runLasso(S, n, id, lambda, W = NULL, mode = 1, update = 100,
  max.gap = 0.5, max.iter = 100, min.t = .Machine$double.eps)
}
\arguments{
\item{S}{sample covariance matrix}

\item{n}{number of samples}

\item{id}{vector assigning variables to nodes}

\item{lambda}{penalty tuning parameter}

\item{W}{optional weight matrix}

\item{mode}{1 (multiattribute; default), 2 (unstructured together), 3 (separately)}

\item{update}{how often to print updates in optimization}

\item{max.gap}{maximum allowable primal/dual gap}

\item{max.iter}{maximum number of iterations to optimize (overrides max.gap)}

\item{min.t}{minimum step size (overrides max.gap, max.iter)}
}
\value{
list of precision, covariance, optimization status, lambda, and number components
}
\description{
Wrapper function for multi-attribute network estimation via lasso
}

