% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lassoWraps.R
\name{runLassoSelect}
\alias{runLassoSelect}
\title{Wrapper function for multi-attribute network estimation plus selection via lasso}
\usage{
runLassoSelect(S, n, id, lambda.range, W = NULL, mode = 1, update = 100,
  max.gap = 0.5, max.iter = 100, min.t = .Machine$double.eps,
  method = "EBIC", Theta.true = NULL, plot = NULL, gamma = 0.5)
}
\arguments{
\item{S}{sample covariance matrix}

\item{n}{number of samples}

\item{id}{vector assigning variables to nodes}

\item{lambda.range}{vector of lambdas to try}

\item{W}{optional weight matrix}

\item{mode}{1 (multiattribute; default), 2 (unstructured together), 3 (separately)}

\item{update}{how often to print updates in optimization}

\item{max.gap}{maximum allowable primal/dual gap}

\item{max.iter}{maximum number of iterations to optimize (overrides max.gap)}

\item{min.t}{minimum step size (overrides max.gap, max.iter}

\item{method}{how to select the best model ('EBIC', 'BIC', 'AIC', or 'hamming')}

\item{Theta.true}{true underlying graph (needed for 'hamming' method only)}

\item{plot}{boolean, whether or not to make diagnostic plot}

\item{gamma}{gamma parameter for EBIC method (default 0.5)}
}
\value{
list of precision, covariance, optimization status, lambda, and number components
}
\description{
Wrapper function for multi-attribute network estimation plus selection via lasso
}

