% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optFunctions.R
\name{updateNodes}
\alias{updateNodes}
\title{Update covariance/precision matrix estimates}
\usage{
updateNodes(t.step, id, S, Omega.tmp, Sigma.tmp, lambda, W)
}
\arguments{
\item{t.step}{step size currently in use}

\item{id}{vector of node identifiers}

\item{S}{sample covariance}

\item{Omega.tmp}{current estimate of precision matrix}

\item{Sigma.tmp}{current estimate of covariance matrix}

\item{lambda}{penalty parameter}

\item{W}{weight matrix}
}
\value{
list of updated covariance and precision estimates
}
\description{
Update covariance/precision matrix estimates
}

