% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optFunctions.R
\name{blockNorms}
\alias{blockNorms}
\title{Get Frobenius norms of submatrices with optional weights}
\usage{
blockNorms(M, id, W = NULL)
}
\arguments{
\item{M}{square matrix of interest}

\item{id}{vector grouping elements of M}

\item{W}{optional weights (square matrix, dimensions equal to length(id))}
}
\value{
matrix of Frobenius norms of submatrices
}
\description{
Get Frobenius norms of submatrices with optional weights
}

