% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optFunctions.R
\name{getEBIC}
\alias{getEBIC}
\title{Get extended BIC of a given network configuration}
\usage{
getEBIC(S, n, Omega, id, gamma = 0.5)
}
\arguments{
\item{S}{sample covariance matrix}

\item{n}{number of samples}

\item{Omega}{estimated precision}

\item{id}{vector grouping elements of S, Omega}

\item{gamma}{EBIC parameter (default 0.5)}
}
\value{
extended Bayesian information criterion for model implied by Omega
}
\description{
In a zero-mean Gaussian graphical model, calculate the extended Bayesian
information criterion (EBIC; Chen and Chen, 2008) of a model with the
proposed precision matrix Omega.  The parameter gamma controls the weight
given towards the component of EBIC related to model size.
}
\references{
Chen, J. and Chen, Z. (2008). Extended Bayesian information criteria for
model selection with large model spaces. Biometrika 95, 759-771.
}

