% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optFunctions.R
\name{getHammingDist}
\alias{getHammingDist}
\title{Get Hamming distance of a given network configuration (truth required)}
\usage{
getHammingDist(Omega, Theta, id)
}
\arguments{
\item{Omega}{estimated precision}

\item{Theta}{true network graph of joint nodes}

\item{id}{vector grouping elements of Omega}
}
\value{
Hamming distance between Theta and graph implied by Omega
}
\description{
Get Hamming distance of a given network configuration (truth required)
}

