% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findPerturbations.R
\name{groupCondLRTs}
\alias{groupCondLRTs}
\title{Sequentially test for perturbations at a series of locations}
\usage{
groupCondLRTs(Y, Omega, Sigma, perturb.mat, ret = "stat")
}
\arguments{
\item{Y}{matrix or data frame (rows=subjects, columns=variables)}

\item{Omega}{precision matrix}

\item{Sigma}{covariance matrix}

\item{perturb.mat}{matrix indicating sets of which variables to test for perturbations}

\item{ret}{return statistics ('stat') or p-values ('pval')}
}
\value{
test statistics for perturbations at each of perturb.mat rows
}
\description{
Given a matrix of possible perturbation locations, perform the sequential
testing procedure described by Griffin et al. Return either test statistics
or p-values.
}
\references{
Griffin, P.J., Johnson, W.E., and Kolaczyk, E.K. Detection of multiple
perturbations in multi-omics biological networks' (under review)
}

