% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lassoWraps.R
\name{runLasso}
\alias{runLasso}
\title{Wrapper function for multi-attribute network estimation via lasso}
\usage{
runLasso(S, n, id, lambda, W = NULL, mode = 1, update = 100,
  max.gap = 0.5, max.iter = 100, min.t = .Machine$double.eps)
}
\arguments{
\item{S}{sample covariance matrix}

\item{n}{number of samples}

\item{id}{vector assigning variables to nodes}

\item{lambda}{penalty tuning parameter}

\item{W}{optional weight matrix}

\item{mode}{estimation mode. Can be 1 (multiattribute; default),
2 (unstructured together), or 3 (separately; only valid for 2-attribute data)}

\item{update}{how often to print updates in optimization}

\item{max.gap}{maximum allowable primal/dual gap}

\item{max.iter}{maximum number of iterations to optimize (overrides max.gap)}

\item{min.t}{minimum step size (overrides max.gap, max.iter)}
}
\value{
list of precision, covariance, optimization status, lambda,
and number components
}
\description{
Performs estimation of a zero-mean multi-attribute Gaussian graphical model
as described by Kolar et al (2014) when \code{mode=1}.
}
\details{
If \code{mode=2}, the \code{id} argument is essentially disregarded, and all
rows/columns in \code{S} are treated as if they represent individual nodes.
If \code{mode=3}, the optimization of \code{mode=2} is performed separately
for each attribute type, and a matrix with 0 on all cross-attribute entries
is returned.
}
\references{
Kolar, M., Liu, H., and Xing, E. P. (2014). Graph estimation from
multi-attribute data. The Journal of Machine Learning Research 15, 1713–1750.
}

